replace all reference of float to float64 now works with last version of go

+++ postscript-go/src/cmd/testpostscript.go
@@ -40,12 +40,12 @@ func main() {
 	i := image.NewRGBA(600, 800)
 	gc := draw2d.NewGraphicContext(i)
 	gc.Translate(0, 380)
-	gc.Scale(1,-1)
+	gc.Scale(1, -1)
 	gc.Translate(0, -380)
 	lastTime := time.Nanoseconds()
 	interpreter := postscript.NewInterpreter(gc)
 	interpreter.ExecuteFile("../../test_files/tiger.ps")
 	dt := time.Nanoseconds() - lastTime
-	fmt.Printf("Draw image: %f ms\n", float(dt)*1e-6)
+	fmt.Printf("Draw image: %f ms\n", float64(dt)*1e-6)
 	saveToPngFile("../../TestPostscript.png", i)
 }
